+++ attributes/default.rb
@@ -1,10 +1,12 @@
-default['mcrouter']['src_dir'] = '/opt/mcrouter'
-default['mcrouter']['install_dir'] = "#{node['mcrouter']['src_dir']}/install"
-default['mcrouter']['user'] = 'mcrouter'
-default['mcrouter']['version'] = 'v0.5.0'
 
-default['folly']['src_dir'] = '/opt/folly'
-default['folly']['version'] = 'v0.53.0'
+default['mcrouter']['install_dir'] = '/usr/local/mcrouter'
+default['mcrouter']['user'] = 'mcrouter'
+default['mcrouter']['version-branch'] = 'release-36-0'
+
+# Specific version of double conversion https://github.com/facebook/mcrouter/blob/master/mcrouter/scripts/recipes/folly.sh#L26
+default['mcrouter']['double_conversion_commit_hash'] = 'ea970f69edacf66bd3cba2892be284b76e9599b0'
+default['mcrouter']['folly_commit_hash'] = 'a1abbb7abcb259acbd94d0d0929b79607a8ce806'
+default['mcrouter']['wangle_commit_hash'] = 'a5503c88e1d6799dcfb337caf09834a877790c92'
 
 default['mcrouter']['local_memcached'] = true
 default['memcached']['port'] = 11_811
@@ -15,17 +17,16 @@ default['mcrouter']['cli_opts'] = {
   'config-file' => '/etc/mcrouter/mcrouter.json',
   'async-dir'   => '/var/spool/mcrouter',
   'log-path'    => '/var/log/mcrouter/mcrouter.log',
-  'pid-file'    => '/var/run/mcrouter/mcrouter.pid',
-  'stats-root'  => '/var/mcrouter/stats',
+  'stats-root'  => '/var/mcrouter/stats'
 }
 
 default['mcrouter']['config'] = {
   'pools' => {
     'A' => {
       'servers' => [
-        'localhost:11811',
-      ],
-    },
+        '0.0.0.0:11811'
+      ]
+    }
   },
   'route' => 'PoolRoute|A',
 }