+++ recipes/_user.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: mcrouter
-# Recipe:: default
+# Recipe:: _user
 #
 # Copyright 2015 EverTrue, Inc.
 #
@@ -15,14 +15,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-include_recipe 'mcrouter::_user'
-include_recipe 'mcrouter::_deps_centos' if node['platform'] == 'centos'
-include_recipe 'mcrouter::_deps_ubuntu_16.04' if node['platform'] == 'ubuntu' && node['platform_version'].to_f	== 16.04
-include_recipe 'memcached::default' if node['mcrouter']['local_memcached']
-include_recipe 'mcrouter::_double_conversion'
-include_recipe 'mcrouter::_folly'
-include_recipe 'mcrouter::_wangle'
-include_recipe 'mcrouter::_install'
-include_recipe 'mcrouter::configure'
-include_recipe 'mcrouter::service'
+user node['mcrouter']['user'] do
+  system true
+  shell '/bin/false'
+end