+++ spec/spec_helper.rb
@@ -1,16 +1,13 @@
 require 'chefspec'
 require 'chefspec/berkshelf'
-require 'coveralls'
 
-# Report coverage to Coveralls
-Coveralls.wear!
+ChefSpec::Coverage.start!
 
 # Generate a report
 ChefSpec::Coverage.start!
 
 RSpec.configure do |config|
-  config.formatter = :documentation
-  config.color = true
-  config.platform = 'ubuntu'
-  config.version = '14.04'
+  config.color      = true            # Use color in STDOUT
+  config.formatter  = :documentation  # Use the specified formatter
+  config.log_level  = :error          # Avoid deprecation notice SPAM
 end